stdlib: fix scheduling issues and add comments
The previous design schedules the start of warmup and end of detailed
simulation together, but after switching the core, the end of detailed
simulation is never reached due to it is never scheduled in the
switched core(s). This commit fixed this problem.

This commit also added comments to all changed components.

This commit added option to change the starting core type for the
riscvmatched board swtichable processor.

Change-Id: I97c5ef7555c4add0dd4a7b367c88bfa2d6187bb1