Functionally correct ART model for prefetch and bypass cache part.
Currently, it can schedule and redirect memory packets if cache should
be bypassed.

The prefetcher can schedule request to the next sequential address and
move data from the prefetch buffer to the current buffer.

When a request comes in, it first check if it is in the current
buffer. If it is, then it sees as current buffer hits and return the
data to the cpu port.
If it is not, it checks if it is in the prefetch buffer. If it is, it
can reponse to the cpu port if the prefetch buffer is valid,
otherwise, it checks if the ICache has it. If the ICache has it, then
fall back to the ICache, otherwise, wait for the prefetch request to
be valid.
If the address can't be find in both the current and prefetch buffer,
it falls back to the cache and also update the next prefetch address
to the next sequential address of the current address.

There are more details to this, for example, if the current prefetch
request is fetching from the next sequential address, then the next
prefetch address is set to next of the next sequential address.