mem-cache: add STM32 ART accelerator for instruction
STM32 ART accelerator has a I-cache, D-cache, and an instruction
prefetcher for the flash memory.
This model does not have a D-cache because the D-cache can be modeled
with existing classic noncoherent cache but not the I-cache and the
instruction prefetcher.

The workflow:
When read request arrives, the ART accelerator first checks the
current buffer then the prefetch buffer.
If hits, then returns the data to the CPU port.
Otherwise, check if request is in the I-cache, if it is, then fetch
from I-cache.
If not in I-cache, check if the prefetcher is currently prefetching
the request.
If it is, then wait for the request to come back and return it to the
CPU port, otherwise, trigger cache miss.

Current limitation:
1) Doesn't have mechanism to deal with privileged accesses.
2) Assume there is only one CPU (because this is modeled after
STM32G4 board which only has one CPU)

Note:
1) ART accelerator's I-cache is a SectorTag cache so it is meant to be
used with the SectorTag tag.
2) The block size is 8 byte here. Because gem5 is not supporting 8-byte
cacheline by default, I changed the NoncoherentCache and SectorTag to
force the cache to use block size as 8.